requirements/base.txt
Django==3.2.8
# SQL
psycopg2-binary==2.9.1
django-environ==0.4.5
django-extensions==3.0.7
djangorestframework==3.11.1